--- apps/oxlint/fixtures/cli/disable_eslint_and_unicorn_alias_rules/.oxlintrc-eslint.json
@@ -1,6 +1,0 @@
-{
-  "plugins": [],
-  "rules": {
-    "eslint/no-negated-condition": "off"
-  }
-}
