--- apps/oxlint/fixtures/cli/disable_eslint_and_unicorn_alias_rules/.oxlintrc-unicorn.json
@@ -1,8 +1,0 @@
-{
-  "plugins": [
-    "unicorn"
-  ],
-  "rules": {
-    "unicorn/no-negated-condition": "off"
-  }
-}
