--- apps/oxlint/fixtures/cli/disable_eslint_and_unicorn_alias_rules/test.js
@@ -1,5 +1,0 @@
-if (!condition) {
-  console.log(123)
-} else {
-  console.log(666)
-}
