+++ crates/oxc_linter/src/rules/shared/eslint_unicorn/mod.rs
@@ -1,0 +1,1 @@
+pub mod no_negated_condition;
