+++ crates/oxc_linter/src/rules/shared/mod.rs
@@ -1,3 +1,5 @@
+mod eslint_unicorn;
 mod jest_vitest;
 
+pub use eslint_unicorn::*;
 pub use jest_vitest::*;
