+++ crates/oxc_linter/src/rules/unicorn/no_negated_condition.rs
@@ -1,0 +1,71 @@
+use oxc_macros::declare_oxc_lint;
+
+use crate::{
+    AstNode,
+    context::LintContext,
+    rule::Rule,
+    rules::shared::no_negated_condition::{DOCUMENTATION, run},
+};
+
+#[derive(Debug, Default, Clone)]
+pub struct NoNegatedCondition;
+
+declare_oxc_lint!(
+    NoNegatedCondition,
+    unicorn,
+    pedantic,
+    pending,
+    docs = DOCUMENTATION,
+    version = "0.0.18",
+);
+
+impl Rule for NoNegatedCondition {
+    fn run<'a>(&self, node: &AstNode<'a>, ctx: &LintContext<'a>) {
+        run(node, ctx);
+    }
+}
+
+#[test]
+fn test() {
+    use crate::tester::Tester;
+
+    let pass = vec![
+        r"if (a) {}",
+        r"if (a) {} else {}",
+        r"if (!a) {}",
+        r"if (!a) {} else if (b) {}",
+        r"if (!a) {} else if (b) {} else {}",
+        r"if (a == b) {}",
+        r"if (a == b) {} else {}",
+        r"if (a != b) {}",
+        r"if (a != b) {} else if (b) {}",
+        r"if (a != b) {} else if (b) {} else {}",
+        r"if (a !== b) {}",
+        r"if (a === b) {} else {}",
+        r"a ? b : c",
+    ];
+
+    let fail = vec![
+        r"if (!a) {;} else {;}",
+        r"if (a != b) {;} else {;}",
+        r"if (a !== b) {;} else {;}",
+        r"!a ? b : c",
+        r"a != b ? c : d",
+        r"a !== b ? c : d",
+        r"(( !a )) ? b : c",
+        r"!(( a )) ? b : c",
+        r"if(!(( a ))) b(); else c();",
+        r"if((( !a ))) b(); else c();",
+        r"function a() {return!a ? b : c}",
+        r"function a() {return!(( a )) ? b : c}",
+        r"!a ? b : c ? d : e",
+        r"!a ? b : (( c ? d : e ))",
+        r"if(!a) b(); else c()",
+        r"if(!a) {b()} else {c()}",
+        r"if(!!a) b(); else c();",
+        r"(!!a) ? b() : c();",
+    ];
+
+    Tester::new(NoNegatedCondition::NAME, NoNegatedCondition::PLUGIN, pass, fail)
+        .test_and_snapshot();
+}
