+++ tasks/lint_rules/src/main.mjs
@@ -6,7 +6,6 @@ import { renderMarkdown } from "./markdown-renderer.mjs";
 import {
   createRuleEntries,
   overrideTypeScriptPluginStatusWithEslintPluginStatus as syncTypeScriptPluginStatusWithEslintPluginStatus,
-  syncUnicornPluginStatusWithEslintPluginStatus,
   updateImplementedStatus,
   updateNotSupportedStatus,
   updatePendingFixStatus,
@@ -62,7 +61,6 @@ void (async () => {
   updateNotSupportedStatus(ruleEntries);
   await updatePendingFixStatus(ruleEntries);
   await syncTypeScriptPluginStatusWithEslintPluginStatus(ruleEntries);
-  syncUnicornPluginStatusWithEslintPluginStatus(ruleEntries);
 
   //
   // Render list and update if necessary
